update: API now also returns daily, weekly and monthly expenses made by the user, so it is not necessary anymore to compute them inside the application update: UserEntity now contains those 3 expenses returned by the server major update: GetUser() and GetTransactions() are no longer part of the MainActivity() class. Those methods were moved to the HomeFragment() class for the GetUser() method and TransactionFragment() for the GetTransactions() update: added OnPageChangeListener() for the ViewPager. This listener let us get from the API the user's data (thanks to the getUser method)if he is on the HomeFragment, and user's transactions (thanks to the getUser method) if he is on the TransactionFragment todo: what does connexion status does ?
