+++ app/src/main/java/com/example/mycomifclient/HomeFragment.kt
@@ -6,16 +6,23 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TableRow
 import android.widget.TextView
-import androidx.constraintlayout.widget.ConstraintLayout
+import android.widget.Toast
 import androidx.fragment.app.Fragment
+import com.example.mycomifclient.database.UserDAO
+import com.example.mycomifclient.database.UserEntity
+import com.example.mycomifclient.serverhandling.HTTPServices
+import com.google.gson.JsonElement
+import com.google.gson.JsonObject
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 import java.util.*
 
 /**
  * Implementation of a fragment with main info (Home fragment)
  */
-class HomeFragment : Fragment() {
+class HomeFragment(private var userDAO: UserDAO) : Fragment() {
     private var listener: OnFragmentInteractionListener? = null
 
     private lateinit var firstName: String
@@ -24,13 +31,19 @@ class HomeFragment : Fragment() {
     private lateinit var weekConsos: String
     private lateinit var monthConsos: String
 
+    private lateinit var user: UserEntity
+
+    //TODO: use basic okHttpClient when the API will be put in production
+    private val retrofitHTTPServices = HTTPServices.create(isSafeConnexion = false)
+
     private var balance = 0f
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        toggleViewStatus(View.INVISIBLE)
+        user = userDAO.getFirst()
+        getUser()
         return inflater.inflate(R.layout.fragment_home, container, false)
     }
 
@@ -38,7 +51,7 @@ class HomeFragment : Fragment() {
      * Updates all views of the fragment according to class data retrieved from API
      * @return None
      */
-    fun updateViews(
+    private fun updateViews(
         firstName: String,
         lastName: String,
         balance: Float,
@@ -111,22 +124,71 @@ class HomeFragment : Fragment() {
     }
 
     /**
-     * Display or Hide the progress bar and fragment views according to the current status
-     * @param status Connexion status
+     * Get the user from API
+     * @return None
+     * @see MainActivity.reconnect
+     * @see handleGetUserResponse
+     */
+    fun getUser() {
+        val token = user.token
+        retrofitHTTPServices.getUser("Bearer $token")
+            .enqueue(object : Callback<JsonObject> {
+                override fun onResponse(
+                    call: Call<JsonObject>,
+                    response: Response<JsonObject>
+                ) {
+                    when (response.raw().code()) {
+
+                        200 -> handleGetUserResponse(response.body(), token)
+
+                        401 -> (activity as MainActivity).reconnect()
+
+                        else -> println("Error")
+                    }
+                }
+
+                override fun onFailure(call: Call<JsonObject>, t: Throwable) {
+                    Toast.makeText(activity, "Error: $t", Toast.LENGTH_LONG).show()
+                }
+            })
+    }
+
+    /**
+     * Handle response to the request for getting a specific user and close activity
+     * @param body response body (JsonObject?)
+     * @param token user token (String)
      * @return None
      */
-    fun toggleViewStatus(status: Int) {
-        view?.findViewById<TextView>(R.id.f_home_text_view_user_name)?.visibility = status
-        view?.findViewById<TextView>(R.id.f_home_text_view_balance)?.visibility = status
-        view?.findViewById<TableRow>(R.id.table_row_stats)?.visibility = status
+    private fun handleGetUserResponse(body: JsonObject?, token: String) {
+        if (body != null) {
+            val userReturned = body.getAsJsonObject("user")
+            val userEntity = UserEntity(
+                userReturned.get("id").asInt,
+                removeQuotes(userReturned.get("first_name")),
+                removeQuotes(userReturned.get("last_name")),
+                removeQuotes(userReturned.get("email")),
+                token,
+                userReturned.get("balance").asInt,
+                body.get("expenses_day").asInt,
+                body.get("expenses_week").asInt,
+                body.get("expenses_month").asInt
+            )
 
-        if (status == View.VISIBLE) {
-            view?.findViewById<ConstraintLayout>(R.id.constraint_layout_progress_bar)?.visibility =
-                View.INVISIBLE
-        } else {
-            view?.findViewById<ConstraintLayout>(R.id.constraint_layout_progress_bar)?.visibility =
-                View.VISIBLE
+            userDAO.nukeUserTable()
+            userDAO.insert(userEntity)
+
+            updateViews(
+                userEntity.firstName,
+                userEntity.lastName,
+                userEntity.balance / 100f,
+                "%.2f".format(userEntity.dailyExpenses / 100f),
+                "%.2f".format(userEntity.monthlyExpenses / 100f),
+                "%.2f".format(userEntity.monthlyExpenses / 100f)
+            )
         }
     }
 
+    private fun removeQuotes(item: JsonElement): String {
+        return item.toString().substring(1, item.toString().length - 1)
+    }
 }
